requirements.txt
python-dotenv
flask
flask-sqlalchemy
flask-migrate
flask-cors
flask-jwt-extended